# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfirmationMethod": {
        "enum": [
          "VOICE",
          "BUTTON",
          "CAREGIVER",
          "AUTO_MISSED"
        ],
        "title": "ConfirmationMethod",
        "type": "string"
      },
...
      "DoseConfirmRequest": {
        "additionalProperties": false,
        "properties": {
          "actual_taken_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Taken Time"
          },
          "confirmation_method": {
            "$ref": "#/components/schemas/ConfirmationMethod"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "confirmation_method"
        ],
        "title": "DoseConfirmRequest",
        "type": "object"
      },
      "DoseLogResponse": {
        "properties": {
          "actual_taken_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Taken Time"
          },
          "confirmation_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfirmationMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "medication_id": {
            "format": "uuid",
            "title": "Medication Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "schedule_id": {
            "format": "uuid",
            "title": "Schedule Id",
            "type": "string"
          },
          "scheduled_time": {
            "format": "date-time",
            "title": "Scheduled Time",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DoseStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "medication_id",
          "schedule_id",
          "scheduled_time",
          "actual_taken_time",
          "status",
          "confirmation_method",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "DoseLogResponse",
        "type": "object"
      },
      "DoseStatus": {
        "enum": [
          "PENDING",
          "TAKEN_ON_TIME",
          "TAKEN_LATE",
          "SKIPPED",
          "MISSED",
          "NEEDS_HELP",
          "VERIFICATION_FAILED"
        ],
        "title": "DoseStatus",
        "type": "string"
      },
      "DoseStatusRequest": {
        "additionalProperties": false,
        "properties": {
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "DoseStatusRequest",
        "type": "object"
      },
      "ErrorResponse": {
        "$defs": {
          "ErrorDetail": {
            "properties": {
              "code": {
                "examples": [
                  "validation_error"
                ],
                "title": "Code",
                "type": "string"
              },
              "details": {
                "default": null,
                "title": "Details"
              },
              "message": {
                "examples": [
                  "Request validation failed"
                ],
                "title": "Message",
                "type": "string"
              },
              "request_id": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "default": null,
                "examples": [
                  "01J00000000000000000000000"
                ],
                "title": "Request Id"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "title": "ErrorDetail",
            "type": "object"
          }
        },
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MedicationForm": {
        "enum": [
          "TABLET",
          "CAPSULE",
          "LIQUID",
          "INJECTION",
          "OTHER"
        ],
        "title": "MedicationForm",
        "type": "string"
      },
      "MedicationResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "brand_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Name"
          },
          "confirmed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed By"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dosage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dosage"
          },
          "form": {
            "$ref": "#/components/schemas/MedicationForm"
          },
          "generic_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generic Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "imprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imprint"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "label_image_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label Image Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pill_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pill Color"
          },
          "pill_image_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pill Image Id"
          },
          "pill_shape": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pill Shape"
          },
          "source": {
            "$ref": "#/components/schemas/MedicationSource"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "with_food": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "With Food"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "generic_name",
          "brand_name",
          "dosage",
          "form",
          "instructions",
          "with_food",
          "pill_color",
          "pill_shape",
          "imprint",
          "label_image_id",
          "pill_image_id",
          "source",
          "confirmed_by",
          "active",
          "created_at",
          "updated_at"
        ],
        "title": "MedicationResponse",
        "type": "object"
      },
      "MedicationSource": {
        "enum": [
          "MANUAL",
          "OCR",
          "CAREGIVER"
        ],
        "title": "MedicationSource",
        "type": "string"
      },
...
      "PrivacyExportResponse": {
        "properties": {
          "ai_verification_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Ai Verification Events",
            "type": "array"
          },
          "caregiver_links": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Caregiver Links",
            "type": "array"
          },
          "devices": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Devices",
            "type": "array"
          },
          "dose_logs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Dose Logs",
            "type": "array"
          },
          "images": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Images",
            "type": "array"
          },
          "medications": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Medications",
            "type": "array"
          },
          "notification_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Notification Events",
            "type": "array"
          },
          "schedules": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Schedules",
            "type": "array"
          },
          "user": {
            "additionalProperties": true,
            "title": "User",
            "type": "object"
          }
        },
        "required": [
          "user"
        ],
        "title": "PrivacyExportResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Medication assistance API. It does not provide medical advice.",
    "summary": "Production API contract for MedGuide AI.",
    "title": "MedGuide AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/caregivers/patients/{patient_id}/dose-logs": {
      "get": {
        "operationId": "get_linked_patient_dose_logs_api_v1_caregivers_patients__patient_id__dose_logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DoseLogResponse"
                  },
                  "title": "Response Get Linked Patient Dose Logs Api V1 Caregivers Patients  Patient Id  Dose Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Linked Patient Dose Logs",
        "tags": [
          "caregivers"
        ]
      }
    },
...
    "/api/v1/dose-logs": {
      "get": {
        "operationId": "list_dose_logs_api_v1_dose_logs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DoseLogResponse"
                  },
                  "title": "Response List Dose Logs Api V1 Dose Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Dose Logs",
        "tags": [
          "dose logs"
        ]
      }
    },
    "/api/v1/dose-logs/today": {
      "get": {
        "operationId": "list_todays_dose_logs_api_v1_dose_logs_today_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DoseLogResponse"
                  },
                  "title": "Response List Todays Dose Logs Api V1 Dose Logs Today Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Todays Dose Logs",
        "tags": [
          "dose logs"
        ]
      }
    },
    "/api/v1/dose-logs/{dose_log_id}/confirm": {
      "patch": {
        "operationId": "confirm_dose_api_v1_dose_logs__dose_log_id__confirm_patch",
        "parameters": [
          {
            "in": "path",
            "name": "dose_log_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dose Log Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoseConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoseLogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Dose",
        "tags": [
          "dose logs"
        ]
      }
    },
    "/api/v1/dose-logs/{dose_log_id}/needs-help": {
      "patch": {
        "operationId": "mark_dose_needs_help_api_v1_dose_logs__dose_log_id__needs_help_patch",
        "parameters": [
          {
            "in": "path",
            "name": "dose_log_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dose Log Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoseStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoseLogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Dose Needs Help",
        "tags": [
          "dose logs"
        ]
      }
    },
    "/api/v1/dose-logs/{dose_log_id}/skip": {
      "patch": {
        "operationId": "skip_dose_api_v1_dose_logs__dose_log_id__skip_patch",
        "parameters": [
          {
            "in": "path",
            "name": "dose_log_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dose Log Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoseStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoseLogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Skip Dose",
        "tags": [
          "dose logs"
        ]
      }
    },
    "/api/v1/medications": {
      "get": {
        "operationId": "list_medications_api_v1_medications_get",
        "parameters": [
          {
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MedicationResponse"
                  },
                  "title": "Response List Medications Api V1 Medications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Medications",
        "tags": [
          "medications"
        ]
      },
      "post": {
        "operationId": "create_medication_api_v1_medications_post",
        "parameters": [
          {
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Medication",
        "tags": [
          "medications"
        ]
      }
    },
    "/api/v1/medications/{medication_id}": {
      "delete": {
        "operationId": "delete_medication_api_v1_medications__medication_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "medication_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Medication Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Medication",
        "tags": [
          "medications"
        ]
      },
      "get": {
        "operationId": "get_medication_api_v1_medications__medication_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "medication_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Medication Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Medication",
        "tags": [
          "medications"
        ]
      },
      "patch": {
        "operationId": "update_medication_api_v1_medications__medication_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "medication_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Medication Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicationUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Medication",
        "tags": [
          "medications"
        ]
      }
    },
    "/api/v1/privacy/export": {
      "get": {
        "operationId": "export_privacy_data_api_v1_privacy_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyExportResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Privacy Data",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/privacy/images/{image_id}": {
      "delete": {
        "operationId": "delete_privacy_image_api_v1_privacy_images__image_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Privacy Image",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/privacy/revoke-caregiver/{link_id}": {
      "post": {
        "operationId": "revoke_caregiver_access_api_v1_privacy_revoke_caregiver__link_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "link_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Link Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Caregiver Access",
        "tags": [
          "privacy"
        ]
      }
    },
...
  },
...
}